+++ opentelemetry-example/src/main/scala/zio/telemetry/opentelemetry/example/BackendApp.scala
@@ -2,10 +2,10 @@ package zio.telemetry.opentelemetry.example
 
 import zio.config.magnolia._
 import zio.config.typesafe.TypesafeConfig
-import zio.telemetry.opentelemetry.Tracing
 import zio.telemetry.opentelemetry.example.config.AppConfig
 import zio.telemetry.opentelemetry.example.http.{ BackendHttpApp, BackendHttpServer }
 import zio._
+import zio.telemetry.opentelemetry.tracing.Tracing
 
 object BackendApp extends ZIOAppDefault {
