+++ opentelemetry-example/src/main/scala/zio/telemetry/opentelemetry/example/ProxyApp.scala
@@ -3,10 +3,10 @@ package zio.telemetry.opentelemetry.example
 import sttp.client3.asynchttpclient.zio.AsyncHttpClientZioBackend
 import zio.config.magnolia._
 import zio.config.typesafe.TypesafeConfig
-import zio.telemetry.opentelemetry.Tracing
 import zio.telemetry.opentelemetry.example.config.AppConfig
 import zio.telemetry.opentelemetry.example.http.{ Client, ProxyHttpApp, ProxyHttpServer }
 import zio._
+import zio.telemetry.opentelemetry.tracing.Tracing
 
 object ProxyApp extends ZIOAppDefault {
