+++ opentelemetry/src/main/scala/zio/telemetry/opentelemetry/tracing/PropagatingSupervisor.scala
@@ -1,4 +1,4 @@
-package zio.telemetry.opentelemetry
+package zio.telemetry.opentelemetry.tracing
 
 import io.opentelemetry.api.trace.Span
 import io.opentelemetry.context.Context
